Finally fix problem with date differences. It was a silly mistake on DateToDays, I had entered a variable name incorrectly. How did this take so long and so much work to be fixed?

+++ src/calendarGUI.cpp
@@ -1220,7 +1220,7 @@ void moveEvent(CalendarEvent* event, int pos) {
     event->startdate.month = em;
     event->startdate.year = ey;
     // calculate new end date based on the difference between the old begin date and the new begin date
-    long long int datediff = DateToDays(ey, em, ed) - DateToDays(oldstartdate.year, oldstartdate.month, oldstartdate.day);
+    long int datediff = DateToDays(ey, em, ed) - DateToDays(oldstartdate.year, oldstartdate.month, oldstartdate.day);
     long int odatedays = DateToDays(event->enddate.year, event->enddate.month, event->enddate.day);
     long int newdatedays = odatedays + datediff;
     //update end date with difference:
