+++ src/timeProvider.hpp
@@ -66,5 +66,5 @@ int isTimeValid(int h, int m, int s);
 int isDateValid(int y, int m, int d);
 
 void DaysToDate(long int d, long int* eyear, long int* emonth, long int* eday);
-long long int DateToDays(int y, int m, int d);
+long int DateToDays(int y, int m, int d);
 #endif
